+++ src/features/home/containers/Body.js
@@ -1,13 +1,28 @@
 import React from 'react';
+import { connect } from 'react-redux';
 
 import '../../../App.css';
+import { filterProducts as filterProductsAction } from '../modules';
 import { ItemsView, NavigationMenu } from '../ui';
 
-const Body = () => (
+const Body = ({ filter, filterProducts, productList, products }) => (
   <div className="Body">
     <NavigationMenu />
-    <ItemsView />
+    <ItemsView filter={filter} filterProducts={filterProducts} productList={productList} products={products} />
   </div>
 );
 
-export default Body;
+const mapDispatchToProps = {
+  filterProducts: filterProductsAction
+};
+
+const mapStateToProps = ({ home }) => ({
+  filter: home.filter,
+  productList: home.productList,
+  products: home.products
+});
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(Body);
