+++ src/features/home/containers/Home.js
@@ -1,27 +1,13 @@
 import React from 'react';
-import { connect } from 'react-redux';
-import { Link } from 'react-router-dom';
 
 import { Body, Header } from '../containers';
-import { testRedux as testReduxAction } from '../modules';
 
-const Home = ({ testData, testRedux }) => (
+const Home = () => (
   <div>
     <Header />
     <Body />
   </div>
 );
 
-const mapDispatchToProps = {
-  testRedux: testReduxAction,
-};
-
-const mapStateToProps = ({ home }) => ({
-  testData: home.testData
-});
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(Home);
+export default Home;
