+++ src/features/home/modules/actions.js
@@ -1,3 +1,4 @@
+export const FILTER_PRODUCTS = '@home/FILTER_PRODUCTS';
 export const TEST_REDUX = '@home/TEST_REDUX';
 
 export function testRedux(testData) {
@@ -5,6 +6,15 @@ export function testRedux(testData) {
     type: TEST_REDUX,
     payload: {
       testData
+    }
+  };
+}
+
+export function filterProducts(filterSymbol) {
+  return {
+    type: FILTER_PRODUCTS,
+    payload: {
+      filterSymbol
     }
   };
 }
