+++ src/features/home/ui/ItemsView.js
@@ -5,132 +5,7 @@ import arrow from '../../../../public/arrow.png';
 import arrowInverted from '../../../../public/arrowInverted.png';
 import searchInverted from '../../../../public/searchInverted.png';
 
-const productList = [
-  {
-    color: '#c393c9',
-    name: '',
-    number: 13,
-    symbol: 'C3'
-  },
-  {
-    color: '#49687c',
-    name: '',
-    number: 13,
-    symbol: 'C4'
-  },
-  {
-    color: '#6785a2',
-    name: '',
-    number: 13,
-    symbol: 'C5'
-  },
-  {
-    color: '#816f6e',
-    name: 'Sulphur',
-    number: 13,
-    symbol: 'S'
-  },
-  {
-    color: '#dfbe73',
-    name: 'LPG',
-    number: 13,
-    symbol: 'L'
-  }
-]
-
-const products = [
-  {
-    actual: '280t',
-    allocation: '300t',
-    color: '#c393c9',
-    company: 'GASCO',
-    id: 1123,
-    month: 'Jan, 2016',
-    name: '',
-    nomination: '290t',
-    number: 13,
-    port: 'Ruwais',
-    status: 'Loading',
-    symbol: 'C3'
-  },
-  {
-    actual: '285t',
-    allocation: '280t',
-    color: '#c393c9',
-    company: 'AD GAS',
-    id: 1124,
-    month: 'Jan, 2016',
-    name: '',
-    nomination: '280t',
-    number: 13,
-    port: 'Shab',
-    status: 'Waiting',
-    symbol: 'C3'
-  },
-  {
-    actual: '330t',
-    allocation: '350t',
-    color: '#49687c',
-    company: 'GASCO',
-    id: 1125,
-    month: 'Feb, 2016',
-    name: '',
-    nomination: '340t',
-    number: 13,
-    port: 'Ruwais',
-    status: 'Finished',
-    symbol: 'C4'
-  },
-  {
-    actual: '280t',
-    allocation: '300t',
-    color: '#6785a2',
-    company: 'GASCO',
-    id: 1126,
-    month: 'Feb, 2016',
-    name: '',
-    nomination: '290t',
-    number: 13,
-    port: 'Ruwais',
-    status: 'Waiting',
-    symbol: 'C5'
-  },
-  {
-    actual: '-',
-    allocation: '500t',
-    color: '#dfbe73',
-    company: 'GASCO',
-    id: 1127,
-    month: 'Jan, 2016',
-    name: 'LPG',
-    nomination: '500t',
-    number: 13,
-    port: 'Ruwais',
-    status: 'Draft',
-    symbol: 'L'
-  },
-  {
-    actual: '380t',
-    allocation: '400t',
-    color: '#816f6e',
-    company: 'GASCO',
-    id: 1128,
-    month: 'Jan, 2016',
-    name: 'Sulphur',
-    nomination: '390t',
-    number: 13,
-    port: 'Ruwais',
-    status: 'Loading',
-    symbol: 'S'
-  }
-]
-
-const filter = {
-  name: "Allocation",
-  items: 12
-}
-
-const ItemsView = () => (
+const ItemsView = ({ filter, filterProducts, productList, products }) => (
   <div className="ItemsView">
     <div className="ItemsViewHeader">
       <button className="ItemsViewButton" style={{ marginLeft: "30px" }} type="button">Inbox</button>
@@ -144,7 +19,7 @@ const ItemsView = () => (
     </div>
     <div className="FilterButtons">
       {productList.map(product => (
-        <button className="FilterButton" type="button" key={product.name} style={{ background: product.color }}>
+        <button className="FilterButton" type="button" key={product.symbol} style={{ background: product.color }} onClick={() => filterProducts(product.symbol)}>
           <div className="FilterButtonSymbol">{product.symbol}</div>
           <div className="FilterButtonName">{product.name}</div>
           <div className="FilterButtonNumber">{product.number}</div>
@@ -176,24 +51,45 @@ const ItemsView = () => (
           </tr>
         </thead>
         <tbody align="center">
-        {products.map(product => (
-          <tr className="TableRow" key={product.id}>
-            <td>
-              <button className="TableRowButton" type="button">
-                <img src={arrow} alt="arrow" height="26" width="26" className="ImageArrow" />
-              </button>
-            </td>
-            <td>{product.id}</td>
-            <td style={{background: product.color, color: "#ffffff", width: "50px"}}>{product.symbol}</td>
-            <td>{product.month}</td>
-            <td>{product.allocation}</td>
-            <td>{product.nomination}</td>
-            <td>{product.actual}</td>
-            <td>{product.company}</td>
-            <td>{product.port}</td>
-            <td style={[product.status ===  "Loading" && {color: "red !important"}, product.status ===  "Finished" && {color: "#25948b !important"}]}>{product.status}</td>
-          </tr>
-        ))}
+        {filter.name === "" ?
+          products.map(product => (
+            <tr className="TableRow" key={product.id}>
+              <td>
+                <button className="TableRowButton" type="button">
+                  <img src={arrow} alt="arrow" height="26" width="26" className="ImageArrow" />
+                </button>
+              </td>
+              <td>{product.id}</td>
+              <td style={{background: product.color, color: "#ffffff", width: "50px"}}>{product.symbol}</td>
+              <td>{product.month}</td>
+              <td>{product.allocation}</td>
+              <td>{product.nomination}</td>
+              <td>{product.actual}</td>
+              <td>{product.company}</td>
+              <td>{product.port}</td>
+              <td style={[product.status ===  "Loading" && {color: "red !important"}, product.status ===  "Finished" && {color: "#25948b !important"}]}>{product.status}</td>
+            </tr>
+          ))
+        :
+          products.filter(product => product.symbol === filter.name).map(product => (
+            <tr className="TableRow" key={product.id}>
+              <td>
+                <button className="TableRowButton" type="button">
+                  <img src={arrow} alt="arrow" height="26" width="26" className="ImageArrow" />
+                </button>
+              </td>
+              <td>{product.id}</td>
+              <td style={{background: product.color, color: "#ffffff", width: "50px"}}>{product.symbol}</td>
+              <td>{product.month}</td>
+              <td>{product.allocation}</td>
+              <td>{product.nomination}</td>
+              <td>{product.actual}</td>
+              <td>{product.company}</td>
+              <td>{product.port}</td>
+              <td style={[product.status ===  "Loading" && {color: "red !important"}, product.status ===  "Finished" && {color: "#25948b !important"}]}>{product.status}</td>
+            </tr>
+          ))
+        }
         </tbody>
       </table>
     </div>
